+++ tests/testapp/urls.py
@@ -5,13 +5,13 @@ from django.shortcuts import render
 
 
 def produce_message(request):
-    messages.info(request, 'Default messages framework')
-    return HttpResponse('OK')
+    messages.info(request, "Default messages framework")
+    return HttpResponse("OK")
 
 
 urlpatterns = [
-    url(r'^admin/', admin.site.urls),
-    url(r'^$', lambda request: render(request, 'base.html')),
-    url(r'^produce_message/$', produce_message),
-    url(r'^404/$', lambda request: render(request, '404.html')),
+    url(r"^admin/", admin.site.urls),
+    url(r"^$", lambda request: render(request, "base.html")),
+    url(r"^produce_message/$", produce_message),
+    url(r"^404/$", lambda request: render(request, "404.html")),
 ]