+++ user_messages/apps.py
@@ -4,5 +4,5 @@ from django.utils.translation import ugettext_lazy as _
 
 
 class UserMessagesConfig(AppConfig):
-    name = 'user_messages'
-    verbose_name = capfirst(_('user messages'))
+    name = "user_messages"
+    verbose_name = capfirst(_("user messages"))